Finish second queue
