+++ src/main/java/com/gramevapp/web/controller/ExperimentController.java
@@ -3,8 +3,7 @@ package com.gramevapp.web.controller;
 import com.gramevapp.web.model.*;
 import com.gramevapp.web.repository.GrammarRepository;
 import com.gramevapp.web.service.*;
-import com.gramevapp.web.service.ExperimentRunnerService.RabbitMqExperimentRunnerService;
-import com.gramevapp.web.service.ExperimentRunnerService.ThreadPoolExperimentRunnerService;
+import com.gramevapp.web.service.experimentrunnerservice.RabbitMqExperimentRunnerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
