+++ src/main/java/com/gramevapp/web/service/experimentrunnerservice/RabbitMqExperimentRunnerService.java
@@ -1,8 +1,11 @@
-package com.gramevapp.web.service.ExperimentRunnerService;
+package com.gramevapp.web.service.experimentrunnerservice;
 
 import com.engine.algorithm.CallableExpGramEv;
+import com.engine.algorithm.RunnableExpGramEv;
 import com.gramevapp.web.model.*;
 import com.gramevapp.web.service.*;
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -30,8 +33,8 @@ import java.util.regex.Pattern;
 
 import static com.engine.util.Common.TRAINING_PATH_PROP;
 
-@Service("threadPoolExperimentRunnerService")
-public class ThreadPoolExperimentRunnerService implements ExperimentRunner{
+@Service("rabbitMqExperimentRunnerService")
+public class RabbitMqExperimentRunnerService implements ExperimentRunner {
 
     private ExperimentService experimentService;
     private SaveDBService saveDBService;
@@ -40,29 +43,32 @@ public class ThreadPoolExperimentRunnerService implements ExperimentRunner{
     private Logger logger;
     private RunService runService;
     private boolean executionCancelled;
-    private Map<Long, Future<Void>> runToFuture;
-    private Map<Long, CallableExpGramEv> runToCallable;
-    private Run[] runElementsInExecution;
+    //private Map<Long, Future<Void>> runToFuture;
+    private Map<Long, RunnableExpGramEv> runToRunnable;
+    //private Run[] runElementsInExecution;
 
-    private ExecutorService threadPool;
+    public static ExecutorService threadPool;
+    @Autowired
+    private RabbitTemplate rabbitTemplate;
 
 
 
-    public ThreadPoolExperimentRunnerService(ExperimentService experimentService, SaveDBService saveDBService
+    public RabbitMqExperimentRunnerService(ExperimentService experimentService, SaveDBService saveDBService
             , RunService runService, Map<Long, Future<Void>> runToFuture
             , Map<Long, CallableExpGramEv> runToCallable, GrammarRepository grammarRepository, UserService userService) {
         this.experimentService = experimentService;
         this.saveDBService = saveDBService;
         this.logger = Logger.getLogger(ThreadPoolExperimentRunnerService.class.getName());
         this.runService = runService;
-        this.runToFuture = runToFuture;
-        this.runToCallable = runToCallable;
+        //this.runToFuture = runToFuture;
+        //this.runToCallable = runToCallable;
 
         this.grammarRepository = grammarRepository;
         this.userService = userService;
 
         int numThreads = Runtime.getRuntime().availableProcessors()/2;
-        this.threadPool = Executors.newFixedThreadPool(numThreads);
+
+        threadPool = Executors.newFixedThreadPool(numThreads);
     }
 
     // Constants
@@ -94,11 +100,7 @@ public class ThreadPoolExperimentRunnerService implements ExperimentRunner{
 
     @Override
     public Future<Void> accept(Run run, String propPath, int crossRunIdentifier, String objective, boolean de, Long expId) {
-        try {
-            return threadPool.submit(runExperimentDetailsServiceWorker(run, propPath, crossRunIdentifier, objective, de));
-        } catch (IOException e) {
-            return null;
-        }
+        return null;
     }
 
     @Override
@@ -146,8 +148,8 @@ public class ThreadPoolExperimentRunnerService implements ExperimentRunner{
         Run run = null;
         String propPath;
 
-        runElementsInExecution = new Run[configExpDto.getNumberRuns()];
-        List<Future<Void>> futures = new ArrayList<>();
+        //runElementsInExecution = new Run[configExpDto.getNumberRuns()];
+        //List<Future<Void>> futures = new ArrayList<>();
 
         //check if need to run more runs
         for (int i = 0; i < configExpDto.getNumberRuns(); i++) {
@@ -160,41 +162,24 @@ public class ThreadPoolExperimentRunnerService implements ExperimentRunner{
             propPath = expPropertiesSet(configExpDto, user, expDataType, grammarFilePath);
             // Run experiment in new thread
             int crossRunIdentifier = exp.isCrossExperiment() ? run.getExperimentId().getIdRunList().indexOf(run) + 1 : -1;
-            Future<Void> future = accept(run, propPath, crossRunIdentifier, configExpDto.getObjective(), configExpDto.isDe(), expId);
-            runToFuture.put(runId, future);
-            futures.add(future);
-            runElementsInExecution[i] = run;
+
+            // send to rabbitmq
+            QueueRabbitMqMessage messageToSend = new QueueRabbitMqMessage(
+                    runExperimentDetailsServiceWorker(run, propPath, crossRunIdentifier, configExpDto.getObjective()
+                            , configExpDto.isDe()), expId, runId, "run");
+
+            rabbitTemplate.convertAndSend(MQConfig.EXCHANGE, MQConfig.RUNS_ROUTING_KEY, messageToSend);
+            //runElementsInExecution[i] = run;
         }
         experimentService.saveExperiment(exp);
         executionCancelled = false;
-
-        // Start all runs and handle their exceptions.
-        threadPool.submit(() ->  {
-            int tareasFinalizadas = 0;
-            for (Future<Void> resultFuture : futures) {
-                try {
-                    resultFuture.get();
-                } catch (InterruptedException e) {
-                    logger.warning("Interrupted thread in service worker");
-                    Thread.currentThread().interrupt();
-                } catch (ExecutionException e) {
-                    Run runFinish = runElementsInExecution[tareasFinalizadas];
-                    runFinish.setStatus(Run.Status.FAILED);
-                    runFinish.setExecReport(runFinish.getExecReport() + "\nUncaught exception: " + e);
-                    String warningMsg = "Uncaught exception: " + e;
-                    logger.warning(warningMsg);
-                } finally {
-                    tareasFinalizadas++;
-                }
-            }
-        });
 
         redirectAttrs.addAttribute("id", exp.getId());
         redirectAttrs.addAttribute("loadExperimentButton", "loadExperimentButton");
         return "redirect:/experiment/expRepoSelected";
     }
 
-    public CallableExpGramEv runExperimentDetailsServiceWorker(Run run, String propPath, int crossRunIdentifier, String objective, boolean de) throws IOException {
+    public RunnableExpGramEv runExperimentDetailsServiceWorker(Run run, String propPath, int crossRunIdentifier, String objective, boolean de) throws IOException {
 
         File propertiesFile = new File(propPath);
         Properties properties = new Properties();
@@ -206,11 +191,11 @@ public class ThreadPoolExperimentRunnerService implements ExperimentRunner{
         properties.setProperty(TRAINING_PATH_PROP, propPath);
 
 
-        CallableExpGramEv obj = new CallableExpGramEv(properties, run,
+        RunnableExpGramEv obj = new RunnableExpGramEv(properties, run,
                 experimentService.findExperimentDataTypeById(run.getExperimentId().getDefaultExpDataType()), runService,
                 saveDBService, crossRunIdentifier, objective, de);
 
-        runToCallable.put(run.getId(), obj);
+        //CallablesSubmiter.runToCallable.put(run.getId(), obj);
         return obj;
     }
 
@@ -404,7 +389,7 @@ public class ThreadPoolExperimentRunnerService implements ExperimentRunner{
         });
     }
 
-    public void removeRunsService(Experiment exp) {
+    public void removeRunsService(Experiment exp) { // check later
         List<Run> oldRunList = new ArrayList<>(exp.getIdRunList());
         //remove old run
         for (Run oldRun : oldRunList) {
@@ -545,11 +530,13 @@ public class ThreadPoolExperimentRunnerService implements ExperimentRunner{
             , DiagramDataService diagramDataService) throws InterruptedException {
         Run run = runService.findByRunId(Long.parseLong(runIdStop));
         Long runId = run.getId();
-        runToCallable.get(runId).stopExecution();
-        run.getDiagramData().setStopped(true);
-        diagramDataService.saveDiagram(run.getDiagramData());
-        run.setStatus(Run.Status.STOPPED);
-        runService.saveRun(run);
+        RunnableExpGramEv runnable = runToRunnable.get(runId);
+
+        QueueRabbitMqMessage stopMessage = new QueueRabbitMqMessage(runnable, run.getExperimentId().getId()
+                , runId, "stop");
+
+        rabbitTemplate.convertAndSend(MQConfig.EXCHANGE, MQConfig.RUNS_ROUTING_KEY, stopMessage);
+
 
         if (model != null) {
             model.addAttribute(EXPDETAILS, run.getExperimentId());
@@ -572,14 +559,6 @@ public class ThreadPoolExperimentRunnerService implements ExperimentRunner{
             if(runIt != null) {
                 runIt.setStatus(Run.Status.STOPPED);
                 runService.saveRun(runIt);
-
-                Long runId = runIt.getId();
-                Future<Void> runFuture = runToFuture.get(runId);
-                if(runFuture != null) {
-                    // No se interrumpen los hilos del threadpool
-                    // Solo detenemos el run.
-                    runToCallable.get(runId).stopExecution();
-                }
                 listRunIt.remove();
                 runIt.setExperimentId(null);
                 runIt.setStatus(Run.Status.STOPPED);
