+++ src/main/java/com/gramevapp/web/service/listener/RabbitListenerService.java
@@ -1,8 +1,8 @@
-package com.gramevapp.web.service;
+package com.gramevapp.web.service.listener;
 
 import com.engine.algorithm.RunnableExpGramEv;
 import com.gramevapp.web.model.Run;
-import com.gramevapp.web.repository.GrammarRepository;
+import com.gramevapp.web.service.*;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.stereotype.Component;
@@ -13,28 +13,18 @@ import java.util.logging.Logger;
 @Component
 public class RabbitListenerService {
     private Logger logger;
-    private ExperimentService experimentService;
-    private SaveDBService saveDBService;
     private RunService runService;
-    private GrammarRepository grammarRepository;
-    private UserService userService;
     private DiagramDataService diagramDataService;
-
-   private RabbitTemplate rabbitTemplate;
-
+    private RabbitTemplate rabbitTemplate;
 
 
-    private final String NUM_THREADS = "2";
 
-    public RabbitListenerService(ExperimentService experimentService, SaveDBService saveDBService, RunService runService
-            , GrammarRepository grammarRepository, UserService userService, DiagramDataService diagramDataService
+    private static final String NUM_THREADS = "2";
+
+    public RabbitListenerService(RunService runService, DiagramDataService diagramDataService
             , RabbitTemplate rabbitTemplate) {
         this.logger = Logger.getLogger(RabbitListenerService.class.getName());
-        this.experimentService = experimentService;
-        this.saveDBService = saveDBService;
         this.runService = runService;
-        this.grammarRepository = grammarRepository;
-        this.userService = userService;
         this.diagramDataService = diagramDataService;
         this.rabbitTemplate = rabbitTemplate;
     }
@@ -50,13 +40,7 @@ public class RabbitListenerService {
     private void startRun(Run run, RunnableExpGramEv runnable) {
         try {
             runnable.run();
-        }catch (Exception ex) { // Esto hay que cambiarlo usando el reportListener
-//            run.setStatus(Run.Status.FAILED);
-//            run.setExecReport(run.getExecReport() + "\nUncaught exception: " + ex);
-//            String warningMsg = "Uncaught exception: " + ex;
-//            logger.warning(warningMsg);
-//            runService.saveRun(run);
-
+        }catch (Exception ex) {
             ReportRabbitmqMessage message = new ReportRabbitmqMessage(run, ex, "run-exception");
             rabbitTemplate.convertAndSend(MQConfig.EXCHANGE ,MQConfig.REPORT_ROUTING_KEY, message);
         }
