+++ src/main/java/com/gramevapp/web/service/listener/ReportsListenerService.java
@@ -1,7 +1,10 @@
-package com.gramevapp.web.service;
+package com.gramevapp.web.service.listener;
 
-import com.engine.algorithm.RunnableExpGramEv;
 import com.gramevapp.web.model.Run;
+import com.gramevapp.web.service.MQConfig;
+import com.gramevapp.web.service.ReportRabbitmqMessage;
+import com.gramevapp.web.service.RunService;
+import com.gramevapp.web.service.SaveDBService;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.stereotype.Component;
 
@@ -10,10 +13,12 @@ import java.util.logging.Logger;
 @Component
 public class ReportsListenerService {
     private RunService runService;
+    private SaveDBService saveDBService;
     private Logger logger;
 
-    public ReportsListenerService(RunService userService) {
+    public ReportsListenerService(RunService userService, SaveDBService saveDBService) {
         this.runService = userService;
+        this.saveDBService = saveDBService;
         this.logger = Logger.getLogger(ReportsListenerService.class.getName());
     }
 
@@ -26,16 +31,22 @@ public class ReportsListenerService {
         runService.saveRun(run);
     }
 
+    // Run that finished successfully
+    private void successRun(Run run) {
+        saveDBService.saveRunAsync(run);
+    }
+
+
     @RabbitListener(queues = MQConfig.REPORTS_QUEUE)
     public void listener(ReportRabbitmqMessage message) {
         Run run = message.getRun();
-        Exception ex = message.getException(); // puede fallar, salta un warning
+        Exception ex = message.getException();
         String code = message.getCode();
 
         // Enhanced java switch.
         switch (code) {
             case "run-exception" -> runExceptionThrew(run, ex);
-            //case "stop" -> stopRun(run, elementToRun);
+            case "finish" -> successRun(run);
             default -> logger.warning("Wrong case in ReportsListenerService");
         }
     }
